fix: make angle wrapping defined on arm64
float->uint16_t conversion of negative values is UB: x86 cvttss2si
wraps mod 2^16 while arm64 fcvtzu saturates to 0. Convert through a
signed wide integer first (llround/int64_t) so the narrowing to
uint16_t is a defined modular reduction, preserving the wrap semantics
the tests encode. Fixes cobAtan and SimAngle round-trip failures on
Apple Silicon.

+++ src/rwe/sim/SimAngle.h
@@ -38,7 +38,7 @@ namespace rwe
     inline SimAngle simAngleFromSimScalar(SimScalar s)
     {
         assert(s.value >= 0.0f);
-        return SimAngle(static_cast<uint16_t>(s.value));
+        return SimAngle(static_cast<uint16_t>(static_cast<int64_t>(s.value)));
     }
 
     SimAngle atan(SimScalar v);
